Use SCIFIO_DEVEL for developer config
The bash scripts also now quote the environment variable when testing
whether it is set, which is much safer.

Lastly, the comment documenting the SCIFIO_DEVEL option has been updated
to hopefully state more clearly in exactly which circumstances you might
want to use it. (It may not be only when building from source, but any
time you want to manually set your own CLASSPATH rather than relying on
the scripts to do it for you.)