AbstractReaderCommand: always log status
Removed safeguard for printing no more than once per second. This
resulted in very confusing messages - if an operation took less than a
second it might not print anything, or you would see one print statement
in a series of plane operations. Better to just remove it, as it is only
once per plane.