SCIFIOToolCommands: add API to get name
Added API to get the name for each SCIFIOToolCommand. A default
implementation in the abstract layer will default to using the class
name in lower case to preserve previous behavior.

This will allow tools to be created with arbitrary (and ideally more
informative) class names, without being tied to their display name when
running SCIFIOTools.